Refactor RegisterMachineOperation and add RegisterWorkerMachineOperation
